+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 - v4.1.1-SNAPSHOT
     - Migrated CI test images from CentOS 7 (EOL) to Rocky Linux 8
     - Fixed NPE "The URI scheme of endpointOverride must not be null" happening during file transfer (e.g. PUT) in some use-cases (snowflakedb/snowflake-jdbc#2572)
+    - Fixed connections.toml auto-configuration behaviour (snowflakedb/snowflake-jdbc#2591):
+      - now defaulting to port 443 instead of 80 when neither port nor protocol is specified
+      - config coming from the JDBC connection string are no longer ignored when auto-configuration sourced items also present (when both present, direct connection config takes precedence)
+
 - v4.1.0
     - Added warning about using plain HTTP OAuth endpoints (snowflakedb/snowflake-jdbc#2556).
     - Fix initializing ObjectMapper when DATE_OUTPUT_FORMAT is specified (snowflakedb/snowflake-jdbc#2545).